+++ backend/src/constants/routes.constants.ts
@@ -1,0 +1,6 @@
+export const TASK_ROUTES = {
+  GET_MY_TASKS: `/tasks`,
+  CREATE_TASK: `/tasks`,
+  UPDATE_TASK: `/tasks/:id/`,
+  DELETE_TASK: `/tasks/:id/`,
+};