+++ backend/src/models/tags/types.d.ts
@@ -1,0 +1,8 @@
+import { Model, Document } from "mongoose";
+import { WALL_STATUS } from "../../constants/global.constants";
+
+export interface Tag {
+  name: string;
+  color: string;
+  isArchived: boolean;
+}