+++ backend/src/models/types.d.ts
@@ -1,0 +1,3 @@
+import { Document, PaginateModel } from "mongoose";
+
+export type PaginatedModel<T> = Document & PaginateModel<T>;